chore: refactoring code structure
