+++ app/src/main/AndroidManifest.xml
@@ -16,14 +16,18 @@
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.Anotei"
-            android:windowSoftInputMode="adjustResize">
+            android:theme="@style/Theme.Anotei">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".ui.screens.note.NoteActivity"
+            android:exported="true"
+            android:theme="@style/Theme.Anotei"
+            android:windowSoftInputMode="adjustResize" />
     </application>
 
 </manifest>
