+++ app/src/main/java/br/com/ascence/anotei/MainActivity.kt
@@ -3,56 +3,51 @@ package br.com.ascence.anotei
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.animation.core.Spring
-import androidx.compose.animation.core.spring
-import androidx.compose.animation.slideInHorizontally
-import androidx.compose.animation.slideOutHorizontally
-import androidx.compose.ui.unit.IntOffset
+import androidx.compose.animation.AnimatedContentTransitionScope
+import androidx.compose.animation.core.tween
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import br.com.ascence.anotei.navigation.DASHBOARD_PATH
 import br.com.ascence.anotei.navigation.NOTE_PATH
-import br.com.ascence.anotei.ui.screens.dashboard.Dashboard
+import br.com.ascence.anotei.ui.screens.dashboard.DashboardScreen
 import br.com.ascence.anotei.ui.screens.note.NoteScreen
 import br.com.ascence.anotei.ui.theme.AnoteiTheme
-import br.com.ascence.anotei.ui.theme.darkThemeColors
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            val slideEffect = spring<IntOffset>(dampingRatio = Spring.DampingRatioMediumBouncy)
 
             val controller = rememberNavController()
 
-            AnoteiTheme(darkColors = darkThemeColors()) {
+            AnoteiTheme {
                 NavHost(navController = controller, startDestination = DASHBOARD_PATH) {
-                    composable(DASHBOARD_PATH) { Dashboard(navController = controller) }
+                    composable(DASHBOARD_PATH) { DashboardScreen() }
                     composable(
                         route = NOTE_PATH,
                         enterTransition = {
-                            slideInHorizontally(
-                                initialOffsetX = { 1000 },
-                                animationSpec = slideEffect
+                            slideIntoContainer(
+                                towards = AnimatedContentTransitionScope.SlideDirection.Companion.Left,
+                                animationSpec = tween(500)
                             )
                         },
                         exitTransition = {
-                            slideOutHorizontally(
-                                targetOffsetX = { -1000 },
-                                animationSpec = slideEffect
+                            slideOutOfContainer(
+                                towards = AnimatedContentTransitionScope.SlideDirection.Companion.Left,
+                                animationSpec = tween(500)
                             )
                         },
                         popEnterTransition = {
-                            slideInHorizontally(
-                                initialOffsetX = { -1000 },
-                                animationSpec = slideEffect
+                            slideIntoContainer(
+                                towards = AnimatedContentTransitionScope.SlideDirection.Companion.Right,
+                                animationSpec = tween(500)
                             )
                         },
                         popExitTransition = {
-                            slideOutHorizontally(
-                                targetOffsetX = { 1000 },
-                                animationSpec = slideEffect
+                            slideOutOfContainer(
+                                towards = AnimatedContentTransitionScope.SlideDirection.Companion.Right,
+                                animationSpec = tween(500)
                             )
                         }
                     ) { NoteScreen(navController = controller) }
