+++ app/src/main/java/br/com/ascence/anotei/ui/common/components/noteoptions/NoteOptions.kt
@@ -1,18 +1,13 @@
-package br.com.ascence.anotei.ui.screencomponents.shared.noteoptions
+package br.com.ascence.anotei.ui.common.components.noteoptions
 
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.CheckCircle
 import androidx.compose.material3.BottomAppBar
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.tooling.preview.PreviewParameter
+import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import br.com.ascence.anotei.data.preview.ColorSchemePreviews
 import br.com.ascence.anotei.data.preview.mock.noteOptionsPreview
 import br.com.ascence.anotei.model.NoteOption
@@ -44,36 +39,6 @@ fun NoteOptionsBar(
     )
 }
 
-@Composable
-private fun Option(option: NoteOption) {
-    Box(
-        contentAlignment = Alignment.Center
-    ) {
-        IconButton(onClick = option.action) {
-            Icon(
-                imageVector = option.icon,
-                contentDescription = option.optionContentDescription,
-                tint = option.iconColor()
-            )
-        }
-        option.checkContentDescription?.let {
-            if (option.showBadge) {
-                Icon(
-                    imageVector = Icons.Default.CheckCircle,
-                    contentDescription = option.checkContentDescription,
-                    tint = AnoteiAppTheme.colors.colorScheme.secondary,
-                    modifier = Modifier
-                        .padding(
-                            bottom = AnoteiAppTheme.spaces.xSmall,
-                            end = AnoteiAppTheme.spaces.xSmall
-                        )
-                        .size(AnoteiAppTheme.spaces.small)
-                        .align(Alignment.BottomEnd)
-                )
-            }
-        }
-    }
-}
 
 @Composable
 fun OptionMainAction(
@@ -95,14 +60,24 @@ fun OptionMainAction(
     }
 }
 
+class NoteOptionsBarPreviewProvider : PreviewParameterProvider<NoteOptionsPresentationType> {
+    override val values: Sequence<NoteOptionsPresentationType>
+        get() = sequenceOf(
+            NoteOptionsPresentationType.PREVIEW_MODE,
+            NoteOptionsPresentationType.EDIT_MODE
+        )
+}
+
 @ColorSchemePreviews
 @Composable
-private fun NoteOptionPreviewLight() {
+private fun NoteOptionPreviewLight(
+    @PreviewParameter(NoteOptionsBarPreviewProvider::class) type: NoteOptionsPresentationType,
+) {
     AnoteiTheme {
         NoteOptionsBar(
             options = noteOptionsPreview,
             onFABClick = {},
-            optionType = NoteOptionsPresentationType.PREVIEW_MODE
+            optionType = type
         )
     }
 }
