+++ app/src/main/java/br/com/ascence/anotei/ui/screens/dashboard/components/notecard/NoteCardHeader.kt
@@ -1,4 +1,4 @@
-package br.com.ascence.anotei.ui.screencomponents.notecard
+package br.com.ascence.anotei.ui.screens.dashboard.components.notecard
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
