+++ app/src/main/java/br/com/ascence/anotei/ui/screens/notelist/NotesListState.kt
@@ -1,4 +1,4 @@
-package br.com.ascence.anotei.ui.screencomponents.notes
+package br.com.ascence.anotei.ui.screens.notelist
 
 import br.com.ascence.anotei.model.Note
